useFetchImages hook 분리
